Fixed one surprising bug, and fixed basically all warnings on MSVC

+++ source/common/string.hpp
@@ -4,7 +4,7 @@
 
 // Allocates a string on free list allocator
 inline char *create_fl_string(const char *buffer) {
-    uint32_t string_length = strlen(buffer);
+    uint32_t string_length = (uint32_t)strlen(buffer);
     char *new_buffer = FL_MALLOC(char, string_length + 1);
 
     memcpy(new_buffer, buffer, string_length + 1);
